Fix #9750: Refactor interfaces into openmetadata-spec module
