+++ common/pom.xml
@@ -28,15 +28,10 @@
   <artifactId>common</artifactId>
 
   <dependencies>
-    <!-- Required dependency for logging -->
     <dependency>
       <groupId>org.slf4j</groupId>
       <artifactId>slf4j-api</artifactId>
       <version>${slf4j.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>com.fasterxml.jackson.core</groupId>
-      <artifactId>jackson-databind</artifactId>
     </dependency>
     <!-- Required dependency for @Slf4j annotation -->
     <dependency>
@@ -69,6 +64,12 @@
           <artifactId>snakeyaml</artifactId>
         </exclusion>
       </exclusions>
+    </dependency>
+    <dependency>
+      <groupId>io.dropwizard</groupId>
+      <artifactId>dropwizard-jersey</artifactId>
+      <version>2.1.4</version>
+      <scope>compile</scope>
     </dependency>
 
   </dependencies>
