+++ openmetadata-service/src/main/java/org/openmetadata/service/exception/BadRequestException.java
@@ -14,6 +14,7 @@
 package org.openmetadata.service.exception;
 
 import javax.ws.rs.core.Response;
+import org.openmetadata.sdk.exception.WebServiceException;
 
 public final class BadRequestException extends WebServiceException {
   private static final String DEFAULT_MESSAGE = "Bad request.";
