+++ openmetadata-service/src/main/java/org/openmetadata/service/exception/CustomExceptionMessage.java
@@ -1,6 +1,7 @@
 package org.openmetadata.service.exception;
 
 import javax.ws.rs.core.Response;
+import org.openmetadata.sdk.exception.WebServiceException;
 
 public class CustomExceptionMessage extends WebServiceException {
   public CustomExceptionMessage(Response.Status status, String message) {
