+++ openmetadata-service/src/main/java/org/openmetadata/service/exception/UnhandledServerException.java
@@ -14,6 +14,7 @@
 package org.openmetadata.service.exception;
 
 import javax.ws.rs.core.Response;
+import org.openmetadata.sdk.exception.WebServiceException;
 
 public class UnhandledServerException extends WebServiceException {
   private static final String MESSAGE = "An exception with message [%s] was thrown while processing request.";
